requirements.txt
python
openai-whisper
ffmpeg-python
deep-translator
torch
